requirements.txt
langchain
langchain-core
langchain-ollama
pydantic
mysql-connector-python
pandas